feat(coprocessor): bigint chain ID across components
Use a bigint for chain ID in the coprocessor, meaning 63 bits of useful
data can be used (instead of 31 bits in some cases).

Using i64 instead of u64 due to more simple mapping to bigint in the DB.

+++ coprocessor/fhevm-engine/test-harness/src/db_utils.rs
@@ -199,7 +199,7 @@ pub async fn setup_test_user(
 }
 
 pub async fn insert_random_tenant(pool: &PgPool) -> Result<i32, sqlx::Error> {
-    let chain_id: i32 = rand::rng().random_range(1..10000);
+    let chain_id: i64 = rand::rng().random_range(1..10000);
     let key_id_i32: i32 = rand::rng().random_range(1..10000);
 
     let verifying_contract_address: String = rand::rng()
